Create SerializedPropertyExtensions and move method
